+++ contracts/index.ts
@@ -1,4 +1,5 @@
 export * from "./analyticsEvents";
 export * from "./gameCatalog";
+export * from "./progression";
 export * from "./roomMatch";
 export * from "./userProfile";